requirements.txt
Django==6.0.3
PyYAML==6.0.3
PyLaTeX==1.4.2
